Update response models to match API output.  Start database

+++ src/ParityFactory.Weather.Models/OpenWeatherApi/Precipitation.cs
@@ -5,8 +5,9 @@ namespace ParityFactory.Weather.Models.OpenWeatherApi
     public class Precipitation
     {
         [JsonPropertyName("1h")]
-        public short? VolumeInPastHour { get; set; }
+        public float? VolumeInPastHour { get; set; }
+
         [JsonPropertyName("3h")]
-        public short? VolumeInPastThreeHours { get; set; }
+        public float? VolumeInPastThreeHours { get; set; }
     }
 }
